Move splitAtSelection to selection.split.
