+++ src/device/__tests__/splitAtSelection.ts
@@ -1,4 +1,3 @@
-import splitAtSelection from '../splitAtSelection'
 import * as selection from '../selection'
 
 it('split at selection with nested nodes', () => {
@@ -13,7 +12,7 @@ it('split at selection with nested nodes', () => {
   // The rise and| fall of Nikola Tesla. (Cursor should be after `and`)
   selection.set(dummyDiv, { offset: 12 })
 
-  const splitResult = splitAtSelection(dummyDiv)
+  const splitResult = selection.split(dummyDiv)
 
   expect(splitResult).toMatchObject({
     left: 'The <b>rise <i>and</i></b>',
